updateAlbum working :)

+++ src/main/java/parcial/backend/repositories/AlbumRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import parcial.backend.entities.Album;
 
 @Repository
-public interface AlbumRepository extends JpaRepository<Album, Long> {
+public interface AlbumRepository extends JpaRepository<Album, Integer> {
 }
